requirements.txt
opencv-python==4.12.0.88
cvzone==1.6.1
plyer==2.1.0
paho-mqtt==2.1.0
Flask==3.1.2
numpy==2.2.6